* 'develop' of https://github.com/rokwire/illinois-app: (40 commits)
  Handled exceptions when system date time is much behind the current date time [#243].
  Fixed unused class warning.
  v2.4.6+506
  Athletics - All Staff tile link is displayed along with Coaching Staff tile list [#239]
  Wellness - PDF links is not loading in web view [#240]
  v2.4.5 changelog
  v2.4.5+505
  Bug/issue 236 (#237)
  YouTube videos plays when screen locked [#235]
  Upload to S3 bucket
  Removed linkage to flutter_image_compress plugin that is unused.
  version: 2.4.4+504
  version: 2.4.3+503
  Do not edit straightly roles from user data (#229).
  Updated sample value of Config().privacyPolicyUrl (#230).
  v2.4.2+502
  Additional fix for [#221]
  Update CHANGELOG.md
  Update AppFrameworkInfo.plist
  v2.4.1+501
  ...

+++ lib/ui/settings/SettingsPrivacyCenterPanel.dart
@@ -437,8 +437,10 @@ class _SettingsPrivacyCenterPanelState extends State<SettingsPrivacyCenterPanel>
 
   Future<void> _deleteUserData() async{
     Analytics.instance.logAlert(text: "Remove My Information", selection: "Yes");
-    await Auth().deleteUserPiiData();
-    await User().deleteUser();
+    bool piiDeleted = await Auth().deleteUserPiiData();
+    if(piiDeleted) {
+      await User().deleteUser();
+    }
     Auth().logout();
   }
